// src/app/types/typesCursos.ts
export interface Curso {
    image: string;
    alt: string;
    imageIcon: string;
    course: string;
    title: string;
    typeCouse: string;
    id: number;
    nome: string;
    rota: string;
    imagePage: string;
    imagePageAlt: string;
    turno: string;
    professionalProfile: string;
    fieldOfActivity: string;
    matrizEmiep: string;
    btnDowload: string;
    nameAdvisor: string;
    emailAdvisor: string;
    testimonials1: string;
    student1: string;
  }
  